edited the file structure to be structured by code type
