+++ .vscode/arduino.json
@@ -1,0 +1,4 @@
+{
+    "port": "COM5",
+    "board": "arduino:avr:uno"
+}
